Handle trash appropriately in mu4e
mu4e places the trash label on any message that has been trashed,
which makes sense. Fastmail, however, treats that as a sign to flat
out delete that message, which is not ideal. Adjust behavior in mu4e
to not apply that, but instead only move it to the trash mimicing
Fastmail.

Fastmail will then fully delete the message after 14 days mbsync will
be smart enough to expunge anything that has been fully deleted.