Update mu4e config and bring versatrial email into setup
There were a number of updates to this that were required to make the
new email work. Bringing mu to a more recent version means better
future proofing overall. Some things had to change in the context tab
to allow for how mu manages emails.

To keep a unified inbox slight changes needed to be made to maildir
shortcuts to keep that functionality. It works identically to how it
did before. Without I would have to go into each email individually.

* Add in versatrial email setup
* Take advantage of debian package instead locally built package for
mu
* Update mu4e to match the new normal mu
* Adjust some variables
* Take advantage of maildir for match func since that's more accurate anyways.